requirements.txt
langchain
langchain-openai
openai
pydantic
unstructured
pdfplumber
spacy
chromadb
streamlit
python-dotenv
pandas
python-docx
PyMuPDF
tiktoken
google-generativeai
joblib